vcpi: enable ipvlan kernel module (vcpi-612)
This module is needed to have VLan in the LXC environment. Witht his
module we can create a docker-network for vlan via

    docker network create -d ipvlan [...]

The idea behind this is, to have two processes listen on the same port
(13400) on different vlans on the same interface.
A manual to this can be found [here].

[here]: https://docs.docker.com/network/drivers/ipvlan/

Signed-off-by: Silvio Fricke <silvio.fricke@softing.com>